moving to workspaces, splitting tests, other things
