+++ batch/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "rust-kzg-bn254-batch-verification"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+rust-kzg-bn254-primitives = { path = "../primitives", version = "0.1.0" }
+ark-bn254 = "0.5.0"
+ark-ec = { version = "0.5.0", features = ["parallel"] }
+ark-ff = { version = "0.5.0", features = ["parallel"] }
+ark-serialize = "0.5.0"
+
+[dev-dependencies]
+rand = "0.8.5"
+criterion = "0.5"
+lazy_static = "1.5"
+tracing = { version = "0.1.41", features = ["log"] }
+tracing-subscriber = "0.3.19"
+ark-std = { version = "0.5.0", features = ["parallel"] }
+rust-kzg-bn254 = { path = "../kzg", version = "0.1.0" }
