+++ batch/README.md
@@ -1,0 +1,9 @@
+# rust-kzg-bn254-batch-verification
+
+[![Crate](https://img.shields.io/crates/v/rust-kzg-bn254.svg)](https://crates.io/crates/rust-kzg-bn254-batch-verification)
+
+This library offers batch verification functions for KZG.
+
+## Warning & Disclaimer
+
+This code is unaudited and under construction. This is experimental software and is provided on an "as is" and "as available" basis and may not work at all. It should not be used in production.
