+++ kzg/kzg_commitment_diagram.png
