+++ primitives/README.md
@@ -1,0 +1,9 @@
+# rust-kzg-bn254-primitives
+
+[![Crate](https://img.shields.io/crates/v/rust-kzg-bn254.svg)](https://crates.io/crates/rust-kzg-bn254-primitives)
+
+This library offers primitive set of structures and functions for generating and interacting with bn254 KZG commitments and proofs in rust.
+
+## Warning & Disclaimer
+
+This code is unaudited and under construction. This is experimental software and is provided on an "as is" and "as available" basis and may not work at all. It should not be used in production.
