+++ primitives/tests/helpers_test.rs
@@ -1,12 +1,14 @@
+use std::{fs::File, io::{BufRead, BufReader, Read}};
+
 use ark_bn254::{Fq, Fq2, Fr, G1Affine, G1Projective, G2Affine, G2Projective};
 use ark_ff::{PrimeField, UniformRand};
 use ark_std::{str::FromStr, One};
-use rust_kzg_bn254::{
-    consts::{BYTES_PER_FIELD_ELEMENT, SIZE_OF_G1_AFFINE_COMPRESSED},
+use rust_kzg_bn254_primitives::{
+    consts::{BYTES_PER_FIELD_ELEMENT, PRIMITIVE_ROOTS_OF_UNITY, SIZE_OF_G1_AFFINE_COMPRESSED},
     helpers::{
         blob_to_polynomial, convert_by_padding_empty_byte, get_num_element, is_on_curve_g1,
-        is_on_curve_g2, remove_empty_byte_from_padded_bytes_unchecked, set_bytes_canonical,
-        set_bytes_canonical_manual, to_byte_array, to_fr_array,
+        is_on_curve_g2, is_zeroed, remove_empty_byte_from_padded_bytes_unchecked,
+        set_bytes_canonical, set_bytes_canonical_manual, to_byte_array, to_fr_array,
     },
 };
 
@@ -43,11 +45,6 @@ fn test_g2_is_on_curve() {
 // Tests deserialization of data and equivalence.
 #[test]
 fn test_blob_to_polynomial() {
-    use ark_serialize::Read;
-    use std::{
-        fs::File,
-        io::{BufRead, BufReader},
-    };
 
     let file = File::open("tests/test-files/blobs.txt").unwrap();
     let mut reader = BufReader::new(file);
@@ -235,70 +232,109 @@ fn test_convert_by_padding_empty_byte() {
     );
 }
 
-#[cfg(test)]
-mod tests {
-    use rust_kzg_bn254::helpers::is_zeroed;
+#[test]
+fn test_is_zeroed_all_zeroes() {
+    // Case where the first byte and the buffer are all zeroes
+    let first_byte = 0;
+    let buf = vec![0, 0, 0, 0, 0];
+    assert!(is_zeroed(first_byte, buf), "Expected true for all zeroes");
+}
 
-    #[test]
-    fn test_is_zeroed_all_zeroes() {
-        // Case where the first byte and the buffer are all zeroes
-        let first_byte = 0;
-        let buf = vec![0, 0, 0, 0, 0];
-        assert!(is_zeroed(first_byte, buf), "Expected true for all zeroes");
-    }
+#[test]
+fn test_is_zeroed_first_byte_non_zero() {
+    // Case where the first byte is non-zero
+    let first_byte = 1;
+    let buf = vec![0, 0, 0, 0, 0];
+    assert!(
+        !is_zeroed(first_byte, buf),
+        "Expected false when the first byte is non-zero"
+    );
+}
 
-    #[test]
-    fn test_is_zeroed_first_byte_non_zero() {
-        // Case where the first byte is non-zero
-        let first_byte = 1;
-        let buf = vec![0, 0, 0, 0, 0];
-        assert!(
-            !is_zeroed(first_byte, buf),
-            "Expected false when the first byte is non-zero"
-        );
-    }
+#[test]
+fn test_is_zeroed_buffer_non_zero() {
+    // Case where the buffer contains non-zero elements
+    let first_byte = 0;
+    let buf = vec![0, 0, 1, 0, 0];
+    assert!(
+        !is_zeroed(first_byte, buf),
+        "Expected false when the buffer contains non-zero elements"
+    );
+}
 
-    #[test]
-    fn test_is_zeroed_buffer_non_zero() {
-        // Case where the buffer contains non-zero elements
-        let first_byte = 0;
-        let buf = vec![0, 0, 1, 0, 0];
-        assert!(
-            !is_zeroed(first_byte, buf),
-            "Expected false when the buffer contains non-zero elements"
-        );
-    }
+#[test]
+fn test_is_zeroed_first_byte_and_buffer_non_zero() {
+    // Case where both the first byte and buffer contain non-zero elements
+    let first_byte = 1;
+    let buf = vec![0, 1, 0, 0, 0];
+    assert!(
+        !is_zeroed(first_byte, buf),
+        "Expected false when both the first byte and buffer contain non-zero elements"
+    );
+}
 
-    #[test]
-    fn test_is_zeroed_first_byte_and_buffer_non_zero() {
-        // Case where both the first byte and buffer contain non-zero elements
-        let first_byte = 1;
-        let buf = vec![0, 1, 0, 0, 0];
-        assert!(
-            !is_zeroed(first_byte, buf),
-            "Expected false when both the first byte and buffer contain non-zero elements"
-        );
-    }
+#[test]
+fn test_is_zeroed_empty_buffer() {
+    // Case where the buffer is empty but the first byte is zero
+    let first_byte = 0;
+    let buf: Vec<u8> = Vec::new();
+    assert!(
+        is_zeroed(first_byte, buf),
+        "Expected true for an empty buffer with a zero first byte"
+    );
+}
 
-    #[test]
-    fn test_is_zeroed_empty_buffer() {
-        // Case where the buffer is empty but the first byte is zero
-        let first_byte = 0;
-        let buf: Vec<u8> = Vec::new();
-        assert!(
-            is_zeroed(first_byte, buf),
-            "Expected true for an empty buffer with a zero first byte"
-        );
-    }
+#[test]
+fn test_is_zeroed_empty_buffer_non_zero_first_byte() {
+    // Case where the buffer is empty and the first byte is non-zero
+    let first_byte = 1;
+    let buf: Vec<u8> = Vec::new();
+    assert!(
+        !is_zeroed(first_byte, buf),
+        "Expected false for an empty buffer with a non-zero first byte"
+    );
+}
 
-    #[test]
-    fn test_is_zeroed_empty_buffer_non_zero_first_byte() {
-        // Case where the buffer is empty and the first byte is non-zero
-        let first_byte = 1;
-        let buf: Vec<u8> = Vec::new();
-        assert!(
-            !is_zeroed(first_byte, buf),
-            "Expected false for an empty buffer with a non-zero first byte"
-        );
+#[test]
+fn test_primitive_roots_from_bigint_to_fr() {
+    let data: [&str; 29] = [
+        "1",
+        "21888242871839275222246405745257275088548364400416034343698204186575808495616",
+        "21888242871839275217838484774961031246007050428528088939761107053157389710902",
+        "19540430494807482326159819597004422086093766032135589407132600596362845576832",
+        "14940766826517323942636479241147756311199852622225275649687664389641784935947",
+        "4419234939496763621076330863786513495701855246241724391626358375488475697872",
+        "9088801421649573101014283686030284801466796108869023335878462724291607593530",
+        "10359452186428527605436343203440067497552205259388878191021578220384701716497",
+        "3478517300119284901893091970156912948790432420133812234316178878452092729974",
+        "6837567842312086091520287814181175430087169027974246751610506942214842701774",
+        "3161067157621608152362653341354432744960400845131437947728257924963983317266",
+        "1120550406532664055539694724667294622065367841900378087843176726913374367458",
+        "4158865282786404163413953114870269622875596290766033564087307867933865333818",
+        "197302210312744933010843010704445784068657690384188106020011018676818793232",
+        "20619701001583904760601357484951574588621083236087856586626117568842480512645",
+        "20402931748843538985151001264530049874871572933694634836567070693966133783803",
+        "421743594562400382753388642386256516545992082196004333756405989743524594615",
+        "12650941915662020058015862023665998998969191525479888727406889100124684769509",
+        "11699596668367776675346610687704220591435078791727316319397053191800576917728",
+        "15549849457946371566896172786938980432421851627449396898353380550861104573629",
+        "17220337697351015657950521176323262483320249231368149235373741788599650842711",
+        "13536764371732269273912573961853310557438878140379554347802702086337840854307",
+        "12143866164239048021030917283424216263377309185099704096317235600302831912062",
+        "934650972362265999028062457054462628285482693704334323590406443310927365533",
+        "5709868443893258075976348696661355716898495876243883251619397131511003808859",
+        "19200870435978225707111062059747084165650991997241425080699860725083300967194",
+        "7419588552507395652481651088034484897579724952953562618697845598160172257810",
+        "2082940218526944230311718225077035922214683169814847712455127909555749686340",
+        "19103219067921713944291392827692070036145651957329286315305642004821462161904",
+    ];
+    let fr_s = data
+        .iter()
+        .map(|s: &&str| Fr::from_str(*s).unwrap())
+        .collect::<Vec<_>>();
+
+    for i in 0..PRIMITIVE_ROOTS_OF_UNITY.len() {
+        let root_of_unity_at_index = PRIMITIVE_ROOTS_OF_UNITY[i];
+        assert_eq!(root_of_unity_at_index, fr_s[i]);
     }
 }
